backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
sqlparse==0.5.5
tzdata==2026.2
